# public/swagger.yaml
swagger: '2.0'
info:
  version: "1.3.0"
  title: ZXInfo API
  description: |
    Welcome to the world's (probaly) greatest Sinclair API!.

    This ZXDB based API is a tool to help developers everywhere create amazing and incredible web sites and apps using data from the 40-plus years area of Sinclair home computers (and clones).
    
    **Ready** to get started?
    
    This is the short version of what you need to do:
    * Be a good API citizen, read and understand how to use the API
    * Keep in touch: tell us about what you're building and talk to other developers
    * Build cool stuff
    
    The ZXInfo API is also the backend for the web site https://zxinfo.dk
    
    Remember that the ZXInfo API suite and portal is an evolving project, developed by dedicated people in their spare time. To the extent that it is possible, we will try to limit any modifications which cause backwards incompatible changes to applications.
    
    Most communication will occur on the community forum on https://spectrumcomputing.co.uk/forums/
    
  contact:
    name: ZXInfo API development
    email: zxinfo_dev@kolbeck.dk
host: api.zxinfo.dk
basePath: /api/zxinfo
...
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /v2/search:
    get:
      tags:
      - "zxinfo"
      summary: "General search for entries"
      description: |
        Returns a List of entries matching input filters.
        
        The search will match input in the following document properties:
          * fulltitle
          * re-released title
          * publisher
          * author, alias and group
        
        Additional options can be specified to futher narrow down search.
      
      operationId: "searchAndFilter"
      produces:
      - "application/json"
      parameters:
      - name: "query"
        in: query
        description: "something to query for"
        type: string
        required: false
      - name: mode
        in: query
        description: Specifies if returned JSON should be 'compact' or 'full'
        required: false
        default: "compact"
        type: string
      - name: sort
        in: query
        description: Specifies sorting mode. Can be on of title_asc, title_desc, date_asc or date_desc.
        required: false
        default: "date_desc"
        type: string
      - name: "contenttype"
        in: query
        description: "Filter by contenttype: SOFTWARE, HARDWARE, BOOK"
        type: string
        required: false
      - name: "language"
        in: query
        description: "Filter by language: English, Spanish, ..."
        type: string
        required: false
      - name: "year"
        in: query
        description: "Filter by year of release"
        type: string
        required: false
      - name: "genretype"
        in: query
        description: "Filter by genretype, e.g. 'Arcade Game'"
        type: string
        required: false
      - name: "genresubtype"
        in: query
        description: "Filter by genresubtype, e.g. 'Action'"
        type: string
        required: false
      - name: "machinetype"
        in: query
        description: "Filter by machinetype, e.g. 'ZX-Spectrum 48K'"
        type: string
        required: false
      - name: "controls"
        in: query
        description: "Filter by controls, e.g. 'Kempston'"
        type: string
        required: false
      - name: "multiplayermode"
        in: query
        description: "Filter by multiplayermode, e.g. 'Turn based'"
        type: string
        required: false
      - name: "multiplayertype"
        in: query
        description: "Filter by multiplayertype, e.g. 'Versus'"
        type: string
        required: false
      - name: "originalpublication"
        in: query
        description: "Filter by originalpublication, e.g. 'Type-in'"
        type: string
        required: false
      - name: "availability"
        in: query
        description: "Filter by availability, e.g. 'Never released'"
        type: string
        required: false
      - name: "group"
        in: query
        description: "Filter by group. See ZXDB documentation"
        type: string
        required: false
      - name: "groupname"
        in: query
        description: "Filter by groupname. See ZXDB documentation"
        type: string
        required: false
      - name: size
        in: query
        description: Specifies the page size for pagination. If used in combination with offset=random, max value allowed is 10.
        required: true
        type: integer
      - name: offset
        in: query
        description: Specifies the page offset for pagination. If 'random' used in combination with size to fetct random documents.
        required: true
        type: integer
      responses:
        # Response code
        200:
          description: "successful operation, JSON document with details"
...
  /authors/{author-name}/games:
    get:
      summary: "Fetches list of entries filtered by author"
      description: |
        Fetches list of entries from a specific author, including:
        * author.name
        * authors.group
        
      tags:
      - "zxinfo"
      parameters:
        - name: author-name
          in: path
          description: Specifies name of author to lookup
          required: true
          type: string
        - name: size
          in: query
          description: Specifies the page size for pagination
          required: true
          type: integer
        - name: offset
          in: query
          description: Specifies the page offset for pagination
          required: true
          type: integer
        - name: mode
          in: query
          description: Specifies if returned JSON should be 'compact' or 'full'
          required: false
          default: "compact"
          type: string
        - name: sort
          in: query
          description: Specifies sorting mode. Can be on of title_asc, title_desc, date_asc or date_desc.
          required: false
          default: "date_desc"
          type: string
      responses:
        # Response code
        200:
          description: |
            An array of games. Adds *X-Total-Count* to reponse header.